refactor: move `get_datasets_from_collection` in `data_utils` + tweaks
